# Deserialize optimizations