… src/serialization/v2_serializer.rs