On d042145 (master + better benchmarks):

```
test serialization::benchmarks::varint_read_rand          ... bench:  16,685,639 ns/iter (+/- 598,312)
test serialization::benchmarks::varint_read_rand_1_byte   ... bench:   2,196,483 ns/iter (+/- 49,426)
test serialization::benchmarks::varint_read_rand_9_byte   ... bench:  16,646,062 ns/iter (+/- 664,383)

test deserialize_large_dense   ... bench:  20,091,180 ns/iter (+/- 511,062)
test deserialize_large_sparse  ... bench:  18,168,099 ns/iter (+/- 463,694)
test deserialize_medium_dense  ... bench:      90,360 ns/iter (+/- 2,269)
test deserialize_medium_sparse ... bench:      79,285 ns/iter (+/- 995)
test deserialize_small_dense   ... bench:       7,996 ns/iter (+/- 145)
test deserialize_small_sparse  ... bench:       2,596 ns/iter (+/- 101)
test deserialize_tiny_dense    ... bench:       2,225 ns/iter (+/- 131)
test deserialize_tiny_sparse   ... bench:         950 ns/iter (+/- 38)
```

Now:

```
test serialization::benchmarks::varint_read_rand              ... bench:  16,280,846 ns/iter (+/- 465,194)
test serialization::benchmarks::varint_read_rand_1_byte       ... bench:   2,151,960 ns/iter (+/- 523,349)
test serialization::benchmarks::varint_read_rand_9_byte       ... bench:  16,428,241 ns/iter (+/- 1,376,795)
test serialization::benchmarks::varint_read_slice_rand        ... bench:   3,894,007 ns/iter (+/- 122,742)
test serialization::benchmarks::varint_read_slice_rand_1_byte ... bench:     695,675 ns/iter (+/- 6,541)
test serialization::benchmarks::varint_read_slice_rand_9_byte ... bench:   3,801,225 ns/iter (+/- 139,819)

test deserialize_large_dense   ... bench:  17,070,681 ns/iter (+/- 414,028)
test deserialize_large_sparse  ... bench:  16,197,753 ns/iter (+/- 460,545)
test deserialize_medium_dense  ... bench:      28,897 ns/iter (+/- 2,052)
test deserialize_medium_sparse ... bench:      22,117 ns/iter (+/- 747)
test deserialize_small_dense   ... bench:       3,811 ns/iter (+/- 77)
test deserialize_small_sparse  ... bench:       1,059 ns/iter (+/- 39)
test deserialize_tiny_dense    ... bench:       1,124 ns/iter (+/- 45)
test deserialize_tiny_sparse   ... bench:         357 ns/iter (+/- 14)
```

So, big wins on tiny and small histograms, less so on very big ones.

For comparison, I wrote some basic benchmarks for the Java impl:

SimpleSerializationBenchmark.largeSparseDecodeFromByteBuffer  thrpt   15     81.947 ±    7.669  ops/s
(or, about 12.2M ns, vs Rust's 16M)
SimpleSerializationBenchmark.medSparseDecodeFromByteBuffer    thrpt   15  13760.020 ±  724.801  ops/s
(72674ns, vs Rust's 22117)

So, we've got some room to improve on the largest histograms, but still a good 3x on the medium-precision ones that people are likely to use.